+++ src/Plugins/ComplexCore/src/ComplexCore/SurfaceNets/MMGeometryOBJ.cpp
@@ -23,7 +23,7 @@ public:
   , m_labels{0, 0}
   {
   }
-  MMQuad(int vi[4], unsigned short labels[2])
+  MMQuad(int vi[4], int32_t labels[2])
   : m_vertexIndices{vi[0], vi[1], vi[2], vi[3]}
   , m_labels{labels[0], labels[1]}
   {
@@ -34,7 +34,7 @@ public:
     for(int i = 0; i < 4; i++)
       vertexIndices[i] = m_vertexIndices[i];
   }
-  void getLabels(unsigned short labels[2])
+  void getLabels(int32_t labels[2])
   {
     for(int i = 0; i < 2; i++)
       labels[i] = m_labels[i];
@@ -44,7 +44,7 @@ public:
     for(int i = 0; i < 4; i++)
       m_vertexIndices[i] = vertexIndices[i];
   }
-  void setLabels(unsigned short labels[2])
+  void setLabels(int32_t labels[2])
   {
     for(int i = 0; i < 2; i++)
       m_labels[i] = labels[i];
@@ -52,7 +52,7 @@ public:
 
 private:
   int m_vertexIndices[4];
-  unsigned short m_labels[2];
+  int32_t m_labels[2];
 };
 
 //
@@ -73,7 +73,7 @@ MMGeometryOBJ::MMGeometryOBJ(MMSurfaceNet* surfaceNet)
   for(int idxVtx = 0; idxVtx < cellMap->numVertices(); idxVtx++)
   {
     int vertexIndices[4];
-    unsigned short quadLabels[2];
+    int32_t quadLabels[2];
 
     // Back-bottom edge
     if(cellMap->getEdgeQuad(idxVtx, MMCellFlag::Edge::BackBottomEdge, vertexIndices, quadLabels) == true)
@@ -113,7 +113,7 @@ MMGeometryOBJ::OBJData MMGeometryOBJ::objData(int label)
   std::map<int, vtxData> vtxDataMap; // key: vertexIndex, value: vtxData for this vertex
   for(std::vector<MMQuad>::iterator itQuad = m_quads.begin(); itQuad != m_quads.end(); itQuad++)
   {
-    unsigned short quadLabels[2];
+    int32_t quadLabels[2];
     itQuad->getLabels(quadLabels);
     if(label == quadLabels[0] || label == quadLabels[1])
     {
@@ -146,7 +146,7 @@ MMGeometryOBJ::OBJData MMGeometryOBJ::objData(int label)
   for(std::vector<MMQuad>::iterator itQuad = m_quads.begin(); itQuad != m_quads.end(); itQuad++)
   {
     int quadVtxIndices[4];
-    unsigned short quadLabels[2];
+    int32_t quadLabels[2];
     itQuad->getLabels(quadLabels);
     itQuad->getVertexIndices(quadVtxIndices);
     if(label == quadLabels[0] || label == quadLabels[1])
